core/requirements.txt
django~=4.2.0
djangorestframework~=3.17.0
psycopg[binary]~=3.3.0
python-decouple~=3.8
pytest~=8.3.0                                                                                                                                                                  
pytest-django~=4.9.0